refactor: 테이블 내부 컴포넌트 디렉토리 변경, context hook 이름 변경
