+++ src/view/user/setting/StudyList/StatusIcon/index.tsx
@@ -4,7 +4,7 @@ import {
   IcnStatusProgress,
   IcnStatusQueued,
 } from "@/asset/svg";
-import type { Status } from "../../index/type";
+import type { Status } from "@/view/user/index/type";
 
 type StatusIconProps = {
   status: Status;
